requirements.txt
beeai
requests
python-dotenv
streamlit
